pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "market-dashboard"
version = "0.1.0"
description = "Live market dashboard — weighted ETF groups with color-coded daily changes"
requires-python = ">=3.11"
dependencies = [
    "yfinance>=0.2.36",
    "streamlit>=1.36",
    "pyxirr>=0.10",
    "empyrical-reloaded>=0.5.9",
    "pandas>=2.0",
    "numpy>=1.24",
    "pdfplumber>=0.11.9",
    "openpyxl>=3.1.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]

[tool.hatch.build.targets.wheel]
packages = ["src/market_dashboard"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
src = ["src"]
line-length = 99
